requirements.txt
flask-openapi3==2.3.1
pydantic==1.10.12
python-multipart==0.0.5
jsonschema==4.17.3
requests==2.28.2
gunicorn==20.1.0
flask_cors==3.0.10
boto3==1.26.73
bcrypt==3.2.0
pyjwt==1.7.1
Werkzeug==2.0.1
flask==2.1.3
python-dotenv
mongoengine
azure-storage-blob 
azure-identity
pdfkit
xmltodict